Rebrand from SurveyShark.site to Survey Shark
Updated multiple files to reflect the new branding, changing all instances of "SurveyShark.site" to "Survey Shark."
Modified the "Last Updated" date in PrivacyPolicy.razor to "4/21/2025" and added a feedback page option to the contact email.
Adjusted wording in TermsOfService.razor for clarity and consistency.
These changes enhance the overall clarity and alignment with the new branding.