Add CanBeMarkedAsLastQuestion and new question type
Added a computed property `CanBeMarkedAsLastQuestion` to the `QuestionViewModel` class to determine if a question can be marked as the last question. Updated its comment for clarity and removed the `[JsonIgnore]` attribute.

Introduced a new `SelectAllThatApplyQuestionViewModel` class, inheriting from `QuestionViewModel`, with an `Options` property to support "select all that apply" question types.